requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
decorator==4.4.2
Django==4.0.6
django-environ==0.9.0
ffmpeg-python==0.2.0
future==0.18.2
idna==3.3
imageio==2.19.5
imageio-ffmpeg==0.4.7
numpy==1.23.1
Pillow==9.2.0
proglog==0.1.10
pycodestyle==2.8.0
python-magic==0.4.27
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
tqdm==4.64.0
urllib3==1.26.10
